Tech debt: Refactored cla_eligible scenarios and steps, removed cla_in_scope.feature as it is commented out anyway.
